backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
python-jose
passlib[bcrypt]
pydantic[email]
